finish the development of the movie list web application
