+++ movie-list/src/App.css
@@ -2,7 +2,7 @@ h1 {
   text-align: center;
 }
 .App {
-  font-family: sans-serif;
+  font-family: Arial, Helvetica, sans-serif;
   text-align: center;
 }
 
@@ -12,10 +12,10 @@ h1 {
   display: flex;
   justify-content: center;
 }
-.nav-container-menu .nav-container-menu-items {
+/* .nav-container-menu .nav-container-menu-items {
   padding-left: 50px;
   overflow: hidden;
-}
+} */
 .nav-container-menu-items > *{
   float: left;
   padding: 10px;
