+++ movie-list/src/containers/home/home.css
@@ -1,4 +1,8 @@
+.homePage{
+  margin-top: 20px;
+}
+
 header{
     font-size:3ch;
     text-align: center;
-  }
+}
