+++ movie-list/src/containers/movieBlockList/movieBlockList.css
@@ -3,22 +3,49 @@
     box-sizing: border-box;
 }
 
+.blockListPage{
+    margin-top: 20px;
+}
+
 .blockMovieBlock{
     display: flex;
     flex-wrap: wrap;
-    margin-left: 80px;
+    padding: 30px;
 }
 
 .blockedMovie{
-    padding: 10px;
-    width: 350px;
+    width: 290px;
+    margin: 0.5rem;
+    position: relative;
 }
 
-.Button{
+.blockedMovie-Button{
     margin-right: 20px;
     margin-left: 20px
 }
 
-.image{
+.blockedMovie-image{
     max-width: 100%;
+}
+
+.blockedMovie:hover .blockedMovie-genres{
+    opacity: 0.6;
+    width: 100%;
+}
+
+.blockedMovie-genres{
+    opacity: 0;
+    position: absolute;
+    transition: width 1s;
+    width: 100px;
+    top: 0px;
+    left: 0;
+    right: 0;
+}
+
+.blockedMovie-genres span{
+    border: 2px solid white;
+    border-radius: 5px;
+    color: white;
+    font-weight: 600;
 }
