+++ movie-list/src/containers/movieList/movieList.css
@@ -8,11 +8,11 @@
 }
 
 .sortButtons{
-    padding-left: 30px;
+    padding-left: 25px;
 }
 
 .sortButtons Button{
-    margin-right: 20px;
+    margin-right: 5px;
 }
 
 .pageNav{
